+++ ar.cpp
@@ -2009,7 +2009,7 @@ void action_replay_memory_reset (void)
 		action_replay_flag = ACTION_REPLAY_ACTIVE;
 }
 
-uae_u8 *save_hrtmon (int *len, uae_u8 *dstptr)
+uae_u8 *save_hrtmon(size_t *len, uae_u8 *dstptr)
 {
 	uae_u8 *dstbak, *dst;
 
@@ -2088,7 +2088,7 @@ uae_u8 *restore_hrtmon (uae_u8 *src)
 	return src;
 }
 
-uae_u8 *save_action_replay (int *len, uae_u8 *dstptr)
+uae_u8 *save_action_replay(size_t *len, uae_u8 *dstptr)
 {
 	uae_u8 *dstbak, *dst;
